requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pydantic==2.*
# 認証関連
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
# データベース
psycopg2-binary==2.9.9
# メール送信（smtplibは標準ライブラリ）
# smtplib2==0.2.1  # 存在しないパッケージのためコメントアウト
